newFeature/development
There was a img path problem due to slash character. It must be dynamically changed according to operating system. It has been fixed.

there was a bug. tags and languages were not updated after delete or edit. It has also been fixed.

initial tags and languages have been added.

languages and text are now converted into lower characters to keep it consistent.

languages and tags are loaded as soon as the initial page loads